requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
langgraph>=0.0.15
anthropic>=0.7.7
langchain-anthropic>=0.0.4
todoist-api-python>=2.1.3
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
loguru>=0.7.2
langfuse
openai

# Test dependencies
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2
pytest-cov>=4.1.0
pytest-env>=1.1.1
